refactor: info badge types
